+++ example/lib/utils/constants.dart
@@ -1,5 +1,0 @@
-const String pageImageKey = 'pageImage';
-
-final Map<String, dynamic> exampleDefaults = {
-  pageImageKey: null,
-};