+++ example/lib/utils/utils.dart
@@ -1,3 +1,2 @@
-export 'constants.dart';
 export 'ExampleScaffold.dart';
 export 'TitleBar.dart';